examples/llm-observability/pyproject.toml
[project]
name = "llm-observability"
version = "0.1.0"
description = "LLM Observability Demo with Phoenix Tracing and LangChain CLI Chatbot"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-core>=0.3.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "arize-phoenix-otel>=0.6.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]

[project.scripts]
llm-chat = "llm_observability.chat:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_observability"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
